src/requirements.txt
anthropic-bedrock==0.8.0
boto3==1.35.24
botocore==1.35.24
cdlib==0.4.0
FlagEmbedding==1.3.2
huggingface-hub==0.23.5
json2xml==5.0.4
llama-index-core==0.10.53.post1
llama-index-embeddings-bedrock==0.2.1
llama-index-llms-bedrock==0.1.12
llama-index-readers-file==0.1.29
llama-index-readers-web==0.1.22
llama-index-vector-stores-opensearch==0.1.12
lru-dict==1.3.0
opensearch-py==2.6.0
pipe==2.2
protobuf==4.25.3
pynvml==11.5.3
pytest==8.2.2
python-dotenv==1.0.1
sentencepiece==0.2.0
sentence-transformers==3.1.1
smart_open==7.0.4
spacy==3.7.5
torch==2.4.1
transformers==4.44.2
tfidf_matcher==0.3.0
